+++ www/warning_continue.php
@@ -4,6 +4,6 @@ use SimpleSAML\Auth\State;
 use SimpleSAML\Auth\ProcessingChain;
 
 $id = $_REQUEST['StateId'];
-$state = State::loadState($id, 'perun:warningTestSP');
+$state = State::loadState($id, 'perun:warning');
 
 ProcessingChain::resumeProcessing($state);